Revert "Revert "adds validations for get by ids endpoints""
This reverts commit c74d3cc25aaaad73be57056f196939bc045ada6c.

+++ src/main/java/com/featuredoc/models/UserRoleId.java
@@ -12,14 +12,14 @@ import lombok.*;
 @AllArgsConstructor
 public class UserRoleId implements Serializable {
 
-    private Integer roleID;
-    private Integer userID;
+    private Long roleID;
+    private Long userID;
 
-    public Integer getUserID() {
+    public Long getUserID() {
         return userID;
     }
 
-    public Integer getRoleID() {
+    public Long getRoleID() {
         return roleID;
     }
